+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 1.8.2 - 2026-09-07
 
 **Highlights:** Homebrew handoffs now accept formulas that preserve signed binaries with literal `skip_clean` paths.